Implement clustering on initial deploy
This change is an initial implentation on clustering. It only supports
clustering on a single deploy, and the nodes in a cluster must stay
constant throughout it's lifecycle. Thus you are not able to: increase
the size of the cluster, update cluster membership (recreates/stemcell
deploys appear to be fine), and most functionality documented here:
https://coreos.com/etcd/docs/latest/v2/runtime-configuration.html