Support scaling up a cluster
This change allows a user to scale up the cluster size by increasing the
value of `instances:` for the etcd instance group.

In order to make this happen there are a few behavior changes:

- On an initial deploy of the etcd cluster, the user must specify the
  `bootstrap.yml` operation file. This operation file has no effect on a
  cluster that already exists if the members are not changing throughout
  deploys. However it cannot be included when scaling up the cluster.

- The etcd job now uses flags passed on the command line, rather than a
  configuration file. The reason for this is that BPM restricts writing to
  the `/var/vcap/jobs/etcd` directory. ETCD provides runtime configuration
  that must be provided to the etcd binary after adding a member to the
  cluster (cluster-state, cluster, etc).

**Note**: Scaling down the cluster is not supported yet. Doing so will
require manual intervention on the cluster. The reason for this is that
there is no lifecycle hook in BOSH that informs you that a VM is
scheduled for deletion.