Support removing members on scaling down
This change introduces a drain script that will remove the node from the
cluster when it detects that it is being shutdown. Currently the only
way to do this in a drain script is to detect that "persistent_disk" is
going to "0" in the `BOSH_JOB_NEXT_STATE` environment variable.