Support scaling up/down from a single node cluster
This change allows users to scale up and down from a single node
cluster. It was necessary to move away from using `etcdctl member list`
for checking membership in the cluster as a single node will not have an
available cluster to query. Instead we can effectively perform the same
by checking for the presence of `/var/vcap/store/etcd/member` which
would indicate that the node was previously apart of an etcd cluster.