Add available etcd options to etcd job
This change adds most of the etcd options/flags that are exposed by the
etcd server to the etcd job. At the moment they are just a pass through
to the etcd binary. I have not added the proxy settings or exposed the
security settings in a similary manner as mutual auth TLS is enforced
when using this release.