Refactor cluster configuration in startup script
This change refactors the bash startup script so that only the CLUSTER
and CLUSTER_STATE variables are set. It results in a clearer startup
process and allows the cluster flags to be set along with the other
flags.

As a result of this change we are able to alphabetize the flags which
results in a consitent and deterministic ordering.