Support auto-compatction-retention
This change adds support for enabling the `auto-compaction-retention`
flag. This flag was missed as it was grouped with the "clustering"
flags. When this property is set to a value n greater than 0, ETCD will
attempt to perform key space compaction every n hours. By default, this
functionality is disabled.

This change also adds an operation file for enabling
`auto-compaction-retention`.

**Note**: If you perform key space compaction, you may want to run
`etcdctl defrag` on a regular basis to defragment the storage space used
by ETCD. Storage space that is not defragmented will still be accessible
to ETCD, but it will not be made free to the host. See the documentation
for more information:
https://coreos.com/etcd/docs/latest/op-guide/maintenance.html.