Add an optional post-start health check
This change adds an optional post-start script that checks `etcdctl
endpoint health`. It only checks the endpoint health of the local ETCD
member.  The `post-start` will timeout if the ETCD member is not healthy
within 10 minutes. In order to enable the post-start health check, apply
the `deployment/operations/enable-post-start-health-check.yml`
operations file or set `properties.wait_on_member_health` to `true` on
the ETCD job.

This change also adds support for using the `LOCAL=true` environment
variable when calling the helper `etcdctl` script. If this envrionment
variable is set to "true", the `etcdctl` command will only be configured
with the endpoint of the local ETCD member